Explore feature design for eDMA3/eDMA4 support
Introduce the additional TCD registers for the eDMA3 and eDMA4 engines.
We'll dispatch reads and writes here, as needed, when managing channels.
Additionally, introduce the DMA controller registers with specialization
for eDMA3 and eDMA4.

Users turn on eDMA3/eDMA4 support with a feature flag. The channel API
is nearly identical (I took away one method for prototyping
convenience). We change how the channels handle some actions, based on
the underlying DMA implementation. Most actions still work on the common
TCD, so those methods are unchanged.

I'm proposing a conditional runtime dispatch approach to select the DMA
controller implementation. This should only kick in when the eDMA3/eDMA4
feature is on.

I definitely missed some variation points in my last commit; so much for
being moderately confident! The compiler helped me find those
differences.

This builds with either feature. I haven't tested this on hardware. I'm
generally not sure how I feel about the approach, and I would look to
clean some of it up if something like this ends up making it into the
package.