Propagate the eDMA instance through Channel types
Channels now convey, at build time, which eDMA peripheral they're
derived from. Users can check the `DMA_INST` at build time to figure out
if their (also strongly-typed) peripherals are compatible with this kind
of eDMA peripheral. To help with this, we'll introduce a trait in the
next commit.

The `Dma` constructors now automatically convey that new `DMA_INST` for
the user.