Clear channel configuration when transfer finishes
Specifically, we should clear the peripheral MUX value on the 1180,
following the recommendations in the reference manual. Otherwise, the
DMA controller will reject a subsequent peripheral MUX value that
matches the *previous* peripheral MUX value. This check evaluates all
DMA channels, including the one we *previously* used.

We're not warning users when the peripheral MUX value fails to stick. We
could add that later. We'll know when it fails: we could read back the
MUX value we've written to make sure it's what we wrote.

I'm clearing the configuration when the Transfer future drops. Since the
transfer future is part of all the other futures, all other futures
implicitly receive this new behavior.