Add fallible DMA configurations
Return an error if we fail to set a DMA channel configuration. See the
parent commit for more information on when this could happen. I'm not
propagating these errors out of the higher-level APIs. However, I expect
there's little chance for panics; users shouldn't duplicate their
channels or peripherals, so we shouldn't observe re-used mappings.